buildiso: add error handling for img mounts

+++ lib/util-iso-image.sh
@@ -20,16 +20,24 @@ copy_overlay(){
     fi
 }
 
+track_img() {
+    info "mount: [%s]" "$2"
+    mount "$@" && IMG_ACTIVE_MOUNTS=("$2" "${IMG_ACTIVE_MOUNTS[@]}")
+}
+
 mount_img() {
+    IMG_ACTIVE_MOUNTS=()
     mkdir -p "$2"
-    info "mount: [%s]" "$1"
-    mount "$1" "$2"
+    track_img "$1" "$2"
 }
 
 umount_img() {
-    info "umount: [%s]" "$1"
-    umount -d "$1"
-    rm -r "$1"
+    if [[ -n ${IMG_ACTIVE_MOUNTS[@]} ]];then
+        info "umount: [%s]" "${IMG_ACTIVE_MOUNTS[@]}"
+        umount "${IMG_ACTIVE_MOUNTS[@]}"
+        unset IMG_ACTIVE_MOUNTS
+        rm -r "$1"
+    fi
 }
 
 configure_plymouth(){
